ENH: rearranging the project file structure
