+++ mily/core/__init__.py
@@ -1,0 +1,1 @@
+from .core import create_qApp  # noqa: F401
