+++ mily/widgets/__init__.py
@@ -1,0 +1,10 @@
+from .table_interface import (  # noqa: F401
+    MTableItemDelegate, MTableInterfaceView, MTableInterfaceWidget,
+    MFunctionTableInterfaceWidget)
+
+from .widgets import (  # noqa: F401
+    MText, MISpin, MFSpin, MComboBox, MCheckBox, MSelector, MDateTime)
+
+# TODO: The following 'widgets': what is the purpose?
+from .widgets import (  # noqa: F401
+    MetaDataEntry, vstacked_label, hstacked_label)
